[CoreBundle] adds a command to empty workspaces by tag
